+++ subprojects/common/analysis/src/main/java/hu/bme/mit/theta/analysis/algorithm/mdd/MddCex.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package hu.bme.mit.theta.analysis.algorithm.mdd.checker;
+package hu.bme.mit.theta.analysis.algorithm.mdd;
 
 import hu.bme.mit.delta.java.mdd.MddHandle;
 import hu.bme.mit.delta.mdd.MddInterpreter;
