+++ subprojects/common/analysis/src/main/java/hu/bme/mit/theta/analysis/algorithm/mdd/expressionnode/LitExprConverter.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package hu.bme.mit.theta.analysis.algorithm.mdd.symbolicnode.expression;
+package hu.bme.mit.theta.analysis.algorithm.mdd.expressionnode;
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
