+++ subprojects/common/analysis/src/test/java/hu/bme/mit/theta/analysis/algorithm/mdd/MddCheckerTest.java
@@ -16,9 +16,6 @@
 package hu.bme.mit.theta.analysis.algorithm.mdd;
 
 import hu.bme.mit.theta.analysis.algorithm.SafetyResult;
-import hu.bme.mit.theta.analysis.algorithm.mdd.checker.MddCex;
-import hu.bme.mit.theta.analysis.algorithm.mdd.checker.MddChecker;
-import hu.bme.mit.theta.analysis.algorithm.mdd.checker.MddWitness;
 import hu.bme.mit.theta.analysis.expr.ExprAction;
 import hu.bme.mit.theta.common.logging.ConsoleLogger;
 import hu.bme.mit.theta.common.logging.Logger;
