+++ subprojects/frontends/petrinet-frontend/petrinet-analysis/src/main/java/hu/bme/mit/theta/frontend/petrinet/analysis/PtNetTransitionNextStateDescriptor.java
@@ -17,8 +17,8 @@ package hu.bme.mit.theta.frontend.petrinet.analysis;
 
 import hu.bme.mit.delta.collections.IntObjCursor;
 import hu.bme.mit.delta.collections.IntObjMapView;
-import hu.bme.mit.theta.analysis.algorithm.mdd.model.AbstractNextStateDescriptor;
-import hu.bme.mit.theta.analysis.algorithm.mdd.model.StateSpaceInfo;
+import hu.bme.mit.theta.analysis.algorithm.mdd.ansd.AbstractNextStateDescriptor;
+import hu.bme.mit.theta.analysis.algorithm.mdd.ansd.StateSpaceInfo;
 import hu.bme.mit.theta.frontend.petrinet.model.Place;
 import hu.bme.mit.theta.frontend.petrinet.model.Transition;
